automatic-instrumentation/todoui-flask/requirements.txt
setuptools>=70.0.0
Faker>=28.4.1
Flask>=3.0.3
opentelemetry-api>=1.28.0
opentelemetry-distro>=0.49b0
opentelemetry-sdk>=1.28.0
opentelemetry-instrumentation>=0.49b0
opentelemetry-exporter-otlp-proto-grpc>=1.28.0
requests>=2.32.3